update entry interface
